# storage/api-docs/api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Smart_home",
    "description": "Документация API для системы Smart_home. Убедитесь что браузер не захешировал авторизационные токены",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8000",
      "description": "Локальный сервер разработки"
    }
  ],
...
"security": [
  { "bearerAuth": [] }
],
  "paths": {
...
    "/api/vacancies/{id}": {
      "get": {
        "tags": ["Vacancies"],
        "summary": "Получение конкретной вакансии",
        "operationId": "show",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Успешное получение конкретной вакансии",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "vacancy": {
                      "type": "object",
                      "properties": {
                        "vacancy_id": {
                          "type": "integer",
                          "example": 14
                        },
                        "title": {
                          "type": "string",
                          "example": "Бариста"
                        },
                        "description": {
                          "type": "string",
                          "example": "Полное описание вакансии..."
                        },
                        "department": {
                          "type": "string",
                          "example": "Офис-менеджер"
                        },
                        "location": {
                          "type": "string",
                          "example": "В офисе"
                        },
                        "type": {
                          "type": "string",
                          "example": "Полный рабочий день"
                        },
                        "salary": {
                          "type": "integer",
                          "example": 42000
                        },
                        "published_at": {
                          "type": "string",
                          "format": "date-time",
                          "example": "2020-10-10T00:00:00Z"
                        },
                        "contact_email": {
                          "type": "string",
                          "format": "email",
                          "example": "email@example.com"
                        },
                        "contact_phone": {
                          "type": "string",
                          "example": "+7 (123) 456-7890"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
        "401": {
            "description": "Неавторизованный доступ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Unauthenticated"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Вакансия не найдена",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "No query results for model [App\\Models\\Vacancies] {id}"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Vacancies"],
        "summary": "Обновление вакансии",
        "operationId": "update",
        "security": [{"bearerAuth": []}],
        "parameters": [{
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
                "type": "integer"
            }
        }],
        "requestBody": {
            "required": true,
            "content": {
                "application/json": {
                    "schema": {
                        "type": "object",
                        "required": ["title", "description", "department", "location", "type", "salary", "contact_email", "contact_phone"],
                        "properties": {
                            "title": {
                                "type": "string",
                                "example": "Измененное название"
                            },
                            "description": {
                                "type": "string",
                                "example": "Изменённое описание вакансии"
                            },
                            "department": {
                                "type": "string",
                                "example": "Должность"
                            },
                            "location": {
                                "type": "string",
                                "example": "Место работы"
                            },
                            "type": {
                                "type": "string",
                                "example": "Полный рабочий день"
                            },
                            "salary": {
                                "type": "integer",
                                "example": 50000
                            },
                            "contact_email": {
                                "type": "string",
                                "format": "email",
                                "example": "pochta@mail.ru"
                            },
                            "contact_phone": {
                                "type": "string",
                                "exmple": "+7-777-777-7777"
                            }
                        }
                    }
                }
            }
        },
        "responses": {
            "200": {
                "description": "Успешное обновление записи",
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "message": {
                                        "type": "string",
                                        "example": "Вакансия успешно обновлена"
                                    }
                                }
                            }
                        }
                    }
                },
                "401": {
                    "description": "Неавторизованный пользователь",
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "message": {
                                        "type": "string",
                                        "example": "Unauthenticated."
                                    }
                                }
                            }
                        }
                    }
                },
                "403": {
                    "description": "Недостаточно прав",
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "message": {
                                        "type": "string",
                                        "example": "User is not an admin"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "delete": {
            "tags": ["Vacancies"],
            "summary": "Удаление конкретной вакансии",
            "operationId": "destroy",
            "security": [{"bearerAuth": []}],
            "parameters": [{
                "name": "id",
                "in": "path",
                "required": true,
                "schema": {
                "type": "integer"
                }
            }],
            "responses": {
                "200": {
                    "description": "Успешное удаление записи",
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "message": {
                                        "type": "string",
                                        "example": "Вакансия успешно удалена"
                                    }
                                }
                            }
                        }
                    }
                },
                "401": {
                    "description": "Неавторизованный пользователь",
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "message": {
                                        "type": "string",
                                        "example": "Unauthenticated."
                                    }
                                }
                            }
                        }
                    }
                },
                "403": {
                    "description": "Недостаточно прав",
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "message": {
                                        "type": "string",
                                        "example": "User is not an admin"
                                    }
                                }
                            }
                        }
                    }
                }
            }

        }
    }
},
"components": {
  "securitySchemes": {
    "bearerAuth": {
      "type": "http",
      "scheme": "bearer",
      "bearerFormat": "JWT"
    }
  }
}
}